Refactored StepParams to Runner
