+++ src/main/java/io/cosmosoftware/kite/interfaces/Runner.java
@@ -1,9 +1,14 @@
-package io.cosmosoftware.kite.steps;
+package io.cosmosoftware.kite.interfaces;
 
 import io.cosmosoftware.kite.report.KiteLogger;
+import io.cosmosoftware.kite.steps.StepPhase;
 import org.openqa.selenium.WebDriver;
 
-public interface StepParams {
+
+/**
+ * Interface for KITE Framework's TestRunner class.
+ */
+public interface Runner {
 
   WebDriver getWebDriver();
   KiteLogger getLogger();
