+++ src/main/java/io/cosmosoftware/kite/steps/ScreenshotStep.java
@@ -1,15 +1,15 @@
 package io.cosmosoftware.kite.steps;
 
 import io.cosmosoftware.kite.exception.KiteTestException;
+import io.cosmosoftware.kite.interfaces.Runner;
 import io.cosmosoftware.kite.report.Reporter;
-import org.openqa.selenium.WebDriver;
 
 import static io.cosmosoftware.kite.util.ReportUtils.saveScreenshotPNG;
 import static io.cosmosoftware.kite.util.ReportUtils.timestamp;
 
 public class ScreenshotStep extends TestStep {
 
-  public ScreenshotStep(StepParams params) {
+  public ScreenshotStep(Runner params) {
     super(params);
     setStepPhase(StepPhase.ALL);
   }
